Store total counts on FetchingArray when fb returns a "count" attribute
Previously we were accessing the FetchingArray.count method, which was
returning the number of entries currently in the array, rather than the
total number of available entries. When displaying statistics, this
makes our numbers look small and sad.

When building a fetching array we now store a "count" attribute if
present in the total_count attribute.

Change-Id: I5dc8f8a629dba727426d151d08e7b79fcbb9ff90
Reviewed-on: https://gerrit.causes.com/3461
Reviewed-by: Lilia Markham <lilia.markham@causes.com>
Tested-by: Lilia Markham <lilia.markham@causes.com>