Mogli is resilient to malformed Facebook responses
Ok, that title was a little bit of title-baiting. Facebook is happy returning
purely a number as a valid JSON response, while the RFC for JSON says that a
valid JSON object must start with a [ or a { (for more info, see:
http://stackoverflow.com/questions/5034444/can-a-json-start-with). The parser
Mogli uses had been raising an exception because it assumed it would respond to
.has_key? if it wasn't already determined to be a list. This commit specifically
handles the case where the decoded JSON is a Numeric.

Release Notes:
Related Airbrake Error: https://causes.airbrake.io/errors/43980389
Related Tracker story: https://www.pivotaltracker.com/story/show/28059743

Change-Id: I08a6f90972297040a40eb1380d24c6788ceaa2f2
Reviewed-on: https://gerrit.causes.com/5488
Reviewed-by: Lilia Markham <lilia.markham@causes.com>
Tested-by: Adam Derewecki <adam@causes.com>